refactor : api 기준 분리 - client : client supabase 사용 - server : server supabase 사용
