+++ src/components/header/header.tsx
@@ -6,8 +6,8 @@ import Image from "next/image"
 import Link from "next/link"
 import { useEffect, useRef, useState } from "react"
 import { toast } from "sonner"
-import { logOut } from "../../libs/api/user"
-import getUserProfileUrl from "../../libs/api/user/get-user-profile"
+import { logOut } from "../../libs/api/client/user"
+import getUserProfileUrl from "../../libs/api/client/user/get-user-profile"
 import { useUserStore } from "../../store/useUserStore"
 import styles from "./header.module.css"
 import ProfileSkeleton from "./profile/profile-skeleton"
