+++ src/components/login/login-form/login-form.tsx
@@ -9,8 +9,8 @@ import Link from "next/link"
 import { useRouter } from "next/navigation"
 import { Controller, useForm } from "react-hook-form"
 import { toast } from "sonner"
-import { logIn } from "../../../libs/api/user"
-import getUserProfileUrl from "../../../libs/api/user/get-user-profile"
+import { logIn } from "../../../libs/api/client/user"
+import getUserProfileUrl from "../../../libs/api/client/user/get-user-profile"
 import styles from "./login-form.module.css"
 
 interface LoginFormData {
