+++ src/components/update-password/update-password-form/update-password-form.tsx
@@ -8,7 +8,7 @@ import { useRouter } from "next/navigation"
 import { useState } from "react"
 import { Controller, useForm } from "react-hook-form"
 import { toast } from "sonner"
-import { updatePassword } from "../../../libs/api/user"
+import { updatePassword } from "../../../libs/api/client/user"
 import styles from "./update-password-form.module.css"
 
 export const UpdatePasswordForm = () => {
